Fix for Qt6.10.1: Explicitly cast unscoped enums to int

Closes #37

See merge request paraview/qttesting!90

+++ pqAbstractItemViewEventTranslator.cxx
@@ -90,8 +90,8 @@ bool pqAbstractItemViewEventTranslator::translateEvent(QObject* Object, QEvent* 
 
       QString info = QString("%1,%2,%3,%4,%5,%6")
                        .arg(mouseEvent->button())
-                       .arg(mouseEvent->buttons())
-                       .arg(mouseEvent->modifiers())
+                       .arg(static_cast<int>(mouseEvent->buttons()))
+                       .arg(static_cast<int>(mouseEvent->modifiers()))
                        .arg(relPt.x())
                        .arg(relPt.y())
                        .arg(idxStr);
